Remove TaxiDispatcherClient.cs and move all it's logic to Program.cs
